+++ src/util/db/DatabaseAdapter.ts
@@ -1,6 +1,7 @@
 import lowdb from 'lowdb';
 import FileSync from 'lowdb/adapters/FileSync';
 
+import EntranceAdapter from './adapters/EntranceAdapter';
 import IgnoreListAdapter from './adapters/IgnoreListAdapter';
 import SoundAdapter from './adapters/SoundAdapter';
 
@@ -8,6 +9,7 @@ export default class DatabaseAdapter {
   public readonly connection: lowdb.LowdbSync<any>;
   public readonly sounds: SoundAdapter;
   public readonly ignoreList: IgnoreListAdapter;
+  public readonly entrances: EntranceAdapter;
 
   constructor() {
     const adapter = new FileSync('db.json');
@@ -16,9 +18,10 @@ export default class DatabaseAdapter {
 
     this.sounds = new SoundAdapter(this.connection);
     this.ignoreList = new IgnoreListAdapter(this.connection);
+    this.entrances = new EntranceAdapter(this.connection);
   }
 
   private ensureDefaults() {
-    this.connection.defaults({ sounds: [], ignoreList: [] }).write();
+    this.connection.defaults({ sounds: [], ignoreList: [], entrances: {} }).write();
   }
 }